Fixes to Categories
